segmented voice connection and audio player logic into TTS_tools.js and the ready event
