+++ events/message.js
@@ -1,7 +1,7 @@
 const { Events } = require('discord.js');
 const { endpoint, guildId } = require('../config.json');
 const { joinVoiceChannel, createAudioPlayer, AudioPlayerStatus, VoiceConnectionStatus } = require('@discordjs/voice');
-const { generateAudioResource } = require('../library/TTS_tools');
+const { generateAudioResource, audioPlayer } = require('../library/TTS_tools');
 
 // 1242754476914249779
 module.exports = {
@@ -23,6 +23,7 @@ module.exports = {
                         "Content-Type": "application/json"
                     }
                 });
+                
                 // Handle error status code
                 if (response.statusCode == 500) {
                     console.log("Error in API request:", response.error);
@@ -37,62 +38,7 @@ module.exports = {
                     console.log("User is not in a voice channel.")
                     return;
                 }
-
-                // Create a new audio player
-                const audioPlayer = createAudioPlayer();
-                console.log("Audio player created!");
-
-                audioPlayer.on(AudioPlayerStatus.Playing, () => {
-                    console.log('Playing audio!');
-                });
                 
-                // Unsubscribe a few seconds after the player becomes idle.
-                // After that, destroy the connection.
-                audioPlayer.on('stateChange', (oldState, newState) => {
-                    console.log(`State changed from ${oldState.status} to ${newState.status}!`);
-                    if (oldState.status == "playing" && newState.status == "idle") {
-                        console.log('Finished playing audio!');
-                        setTimeout(() => {
-                            subscription.unsubscribe();
-                            voiceConnection.destroy();
-                            audioPlayer.stop();
-                        }, 3000);
-                    }
-                    
-                });
-            
-                audioPlayer.on('error', (e) => {
-                    console.log(`Error: ${e}`);
-                });
-
-                // Establish a Voice Connection
-                const voiceConnection = joinVoiceChannel({
-                    channelId: message.member.voice.channelId, // Alyssa's Den VC
-                    guildId: guildId,
-                    adapterCreator: message.guild.voiceAdapterCreator
-                });
-            
-                // Handle Disconnection
-                voiceConnection.on(VoiceConnectionStatus.Disconnected, async (oldState, newState) => {
-                    try {
-                        await Promise.race([
-                            entersState(connection, VoiceConnectionStatus.Signalling, 5_000),
-                            entersState(connection, VoiceConnectionStatus.Connecting, 5_000),
-                        ]);
-                        // Seems to be reconnecting to a new channel - ignore disconnect
-                    } catch (error) {
-                        // Seems to be a real disconnect which SHOULDN'T be recovered from
-                        connection.destroy();
-                    }
-                });
-
-                // When ready, subscribe to the audio player.
-                voiceConnection.on(VoiceConnectionStatus.Ready, () => {
-                    console.log("Voice Connection Established!");
-                    subscription = voiceConnection.subscribe(audioPlayer);
-                });
-
-                // Generate audio
                 const audioResource = await generateAudioResource(responseJSON.result);
                 audioPlayer.play(audioResource);
